Fix Mumble URL parsing
URLs including password but no username were parsed incorrectly. The
Mumble Wiki's specification is somewhat confusing for this case. The
General URL specification implies an username is needed to use a
password but later on page it is clear that password can be used
without an accompanying username.

The desktop client also supports these URLs.